move storage/utils to kstd/env
